Add HistoryList class to datamodels
The HistoryList class is a container class that golds the list of
history entries. It was written so that datamodels does not handle
history entriies directly, but uses the asdf methods
get_history_entries and add_history_entry instead. List methods needed
for manipulating history entries were written, but some list methods
are unsupported.