white_light should handle segmented data, and data with multiple orders
The white_light step failed because the original file had been segmented and
the code was using input.meta.exposure.nints for the number of integrations.
Furthermore, the input can contain several (up to three) tables for each
integration, i.e. one for each spectral order.  The section of the code that
computes the time for each integration was modified to take into account the
possibility that the input contained more than one spectral order.  The
number of tables for each spectral order is determined from the input tables,
rather than relying on the NINTS keyword.
See issue #2148.

While testing the above change using NIRISS SOSS data, it was noticed that
the extract_1d step was printing a warning message for every integration
(for spectral orders 2 and 3), e.g. "Using RELSENS, 736 elements were
extrapolated; these values will be set to 1."  This should only have been
printed for the first integration (see issue #2079).  This was printed by
function interpolate_response, so a `verbose` argument was added to that
function, and verbose is set to False after the first integration.