Add tests for HistoryList methods
This update adds test_historylist_methods to test_history.py. It tests
all the methods in history.py. There are also a few changes to
history.py and model_base.py. I remove some imports that are no longer
needed, moved the test that the input is a list to a new function,
_iterable, and removed that test from model_base.py.