Update data models and steps for the integration times table.
The extract_1d step was modified to copy the times from the input INT_TIMES
table to metadata for the output spectra, but only for TSO data, and only if
there is an INT_TIMES table in the input.  PR #1987 included changes to
extract_1d, but the current commit fixes some bugs in the earlier version
and adds a check for TSO data.

The ramp_fit step has been modified to copy the input INT_TIMES table to the
output model for integration-specific results for TSO data, and the extract_2d
step has been modified to copy the input INT_TIMES table to the output model
for 3-D NIRSpec NRS_BRIGHTOBJ data.

The tso_photometry and white_light steps have been modified to copy the
times from the INT_TIMES table (if it is present in the input) instead of
computing the times directly.

Some of the data models and schemas were modified to include an int_times
attribute.  Some had been modified earlier, in PR #1987, but further updates
were needed.  The following data models now include an int_times attribute:
Level1bModel, RampModel, CubeModel, SlitModel, and MultiSpecModel.