Skip spectral extraction if dispaxis could not be determined
There's now a default of None for the dispersion axis, primarily so that
the initial value (along with other parameters) can be logged without
crashing.

The function that finds the dispersion axis from the wavelengths uses either
log.warning or log.debug to write the value from the log file and the value
determined from the wavelengths, depending on whether there was a discrepancy
or not.

If the dispersion direction could not be determined, spectral extraction
will be skipped.  Extraction will also be skipped if the spectral order
is 0, unless the data were taken with a prism.

See issue #2102.