For extract_1d, find nod/dither offset from XOFFSET, YOFFSET
Method offset_from_offset was modified to include slit as a new argument.
If slit is defined, some of the parameters (e.g. v2_ref) will now be gotten
from slit; otherwise, those parameters will be gotten from input_model.
Checks were added for the existence of attributes.  For NIRSpec data, if
v3yangle or vparity is not defined in the input, a default value is used.

See issue #1894.