Rewrote the hasattr method in Datamodels
The __hasattr__ method has been replaced by hasattr. It tests to see
if a node has a specified attribute, much like the hasattr function in
Python. However, the hasattr method does not create the attribute
while the hasatter function does. This is a peculiarity of the
implementation of Datamodels nodes and not a general issue for Python.

I added a new test for hasattr and went through the datamodels code to
replace the use of the hasattr function with the new method where that
was appropriate.