Fix handling of default data array name
Datamodels normally store their most significant data in an array
named 'data'. Datamodels allows this to be changed in a subclass by
overriding the method get_primary_array_name. The base class returns
the name 'data'. I've modified it so that it now checks the schema and
returns the empty string if it is not found. I used the empty string
rather than None in case a user feeds the result directly into
getattr. A value of None will throw an error and both are falsey. The
new code prevents datamodels that contain tables from reporting that
they have a data array.

I've modified the shape method/property to test the result of
get_primary_array_name, which restores the behavior I accidentally
changed when adding the hasattr method.

I've modified the data getter/setter in NDModel to call
get_primary_array_name. This fixes a bug where the code assumed the
name was 'data'. NDModel is a mixin class that provides compatibility
with astropy's NDData.

I've added methods overriding get_primary_array_name to AMilgModel and
LinearityModel, since, according to the schema, these two classes,
along with MaskModel, have data arrays with names other than 'data'.