Fix crash in make_header and add tests. Fixes #2279
A user reported make_header was crashing. The problem was caused by
the validation method in datamodels being renamed. To catch similar
problems in the future, test_make_header has been added to the
jwst.datamodels tests. This required putting the bulk of the code in
the make_header script into make_header.py in jwst.datamodels.