Added annotation to ASDF exception message when embedded ASDF fails to load so that end users get more info.
This enables CRDS certify to issue an error message like this:

     Validation error : JWST Data Models: ERROR loading embedded ASDF: newline not found

instead of a less helpful message like this:

     Validation error : JWST Data Models: newline not found

Certify options for trapping this robustly are pretty limited, this particular
pull was inspired by a corrupt .fits with bad/missing ASDF extension which
bubbles up to CRDS as "ValueError('newline not found')".

IMHO the datamodels are probably the right place to add explanation since
they're doing the funky embedded ASDF thing.   I guess an alternative would
be to improve the ASDF embed functions.