Update code to compute primary array name
Datamodels has a method named get_primary_array_name that returns the
name of the primary data array. Its purpose is to allow the name to be
overridden by a subclass, something only a few subclasses do. This
update sets the name to the empty string when a datamodel does not
have a primary data array. It also updates ndmodel so that it calls
get_primary_array_name instead of assuming the name is 'data', as it
did before. It also overrides the method in a few subclasses where the
schema indicated another array was the primary array.