Update utc_to_tdb to also update VELOSYS.
The utc_to_tdb module in jwst.lib has been modified to (optionally) compute
the radial velocity (m / s) of JWST with respect to the solar-system
barycenter, and to assign that value to keyword VELOSYS in the SCI header
of the specified FITS file.  See JWST issue #1742 and JIRA issue JWSTKD-193.