Remove dangling pointers to memory mapped data
When datamodels opens a fits file, it also opens a memory mapped array
to its data. This uses a file handle, which normally is released when
the fits file and its associated memory mapped data is
closed. However, the memory mapped data is not closed if a reference
is held to it. This happens in the current version of datamodels and
this update is my best attempt to close the dangling pointers to the
memory mapped data when the datamodel is closed. The change to
fits_support.py removes a second reference to the data that is created
when casting the data to a new type. The change to model_base.py
removes all references to the data when closing the model. And the
change to validate.py moves the asdf context, which is simply a way to
access the validator, from class scope to function scope.