Remove dangling pointers to data from datamodels (#2360)
An issue was filed (#2282) that datamodels was not closing file
handles after opening them. After some investigation, it looks like
there are two problems. The first problem is that a fits file is
opened, normally the data in the file is accessed as a memory mapped
array. When the fits file is closed, the handle to the memory map is
only closed if there is a single reference to the memory map. Normally
this condition is not met. Instead, the memory containing the fits
descriptor is released, and some time after this the memory is
garbage collected. When this happens, the file handle to the memory
mapped array is closed. This causes a problem when a reference to the
fits file descriptor is not released. In that case, the file handle is
not closed and if one is looping over a set of fits files, one file
handle will be leaked for each file.

This fix contains a partial fix for the problem in that it closes some
extra references to the memory mapped data. However, since I have not
located them all, the leak is not yet fixed. The first change was to
add a new method to model_base.py, _drop_arrays, which walks the data
structure and deletes all numpy arrays. The second change is in
from_fits_hdu in fits_support.py, where instead of creating a new copy
of a numpy array before casting it to a new type, it over writes the
original array. I also added an explicit close to the asdf object that holds the validator
context, so that any attributes it holds, including the memory mapped
buffer, can be reaped.