Add warning message to datamodels.open when model_type not found [#2248] (#2401)
The generic open function in datamodels first uses the metadata
parameter model_type to decide what class to instantiate a model
as. If it cannot find it, datamodels falls back on other information
to guess the type. If it has to guess, it now issues a warning message
with the name of the file (if found) and the class has guessed:

NoTypeWarning: model_type not found. Opening flat.fits as a ImageModel