Move no model type warning in datamodels
The recently introduced warning that is printed when the DATAMODL
keyword is not present in a FITS file is a subtle change to the
existing interface. This is because it precedes any warning produced
when a FITS file is opened and confuses any code that might be
trapping warning messages when a datamodel is opened. This update
preserves the existing interface by moving the test to after the
datamodel is opened, so any error or warning during the open takes
precedence over it. This update changes nothing Other than changing
the location of the warning in the output.

This change also adds a line about the warning to the CHANGES file.