Fix problems in schema editor uncovered by flake8 (#2394)
Flake8 found some small problems in the schema editor that needed to
be fixed.  The editor used to track if a schema field had been
modified with the boolean flag done. This flag is no longer used and
was removed. Model schema field names sometimes differ from the
equivalent keyword database field name. One line where the keyword
database name was used instad of the model schema name was
changed. And datatype names can differ between the model schema and
keyword database. The code was not catching that integer and int are
the same, but now it is.