Fix bug in resample_spec wavelength reverse lookup table (#2422)
When resample_spec creates an output frame, the wavelengths are done via tabular interpolation. Both the forward transform (pixels to wavelength) and reverse transform (wavelength to pixels) need to be defined.

This fixes a bug in the reverse where for MIRI LRS fixed slit data, the wavelength is descending as a function of pixel index. For scipy.interpolate, which is the underlying interpolation method used in the tabular models, the interpolation points need to be strictly ascending.