Remove warning message when opening a model from a shape or array
Datamodels currently issues a warning message when a new model is
opened if that model does not have a DATAMODL header keyword. The code
will issue this message even when a model is created from scratch using
a shape or numpy array. This update only issues a warning message if the
datamodel is opened from a file or hdu with an associated file.

It also suppresses several warning messages in tests about the missing
DATAMODL header keyword.