Improve the code for finding the dispersion direction
Function find_dispaxis was added.  This determines the dispersion direction
by comparing the rate at which wavelength changes from pixel to pixel, in
the horizontal and vertical directions.  All pixels with non-NaN and
non-zero values of wavelength are now used, rather than just three points
near the middle of the image.

The function that combines the flat-field images (slowly varying with
wavelength) with the tabular flat field (varying more quickly with wavelength)
replaces the NaN or zero wavelength values with an average wavelength, to
reduce interpolation problems near the edges of the bounding box.

See issue #2487.