Initialize data arrays and tables from function args in model_base (#2502)
The classes in datamodels contain a lot of code to initialize arrays
from function arguments. This update centralizes this code in
DataModel, the base class of the pther classes. It reads the arguments
from **kwargs, sees if the arguments are mentioned in the schema, and
if so, uses the value of the argument to initialize an array or table.
The intialization of the arrays in each of the subclasses has been
removed. This has no effect on the code that calls the DataModel
subclasses, but any programmer adding a new subclass should examine
the code first. Eliminating the initialization code has allowed many
of the subclasses to eliminate their \_\_init\_\_ function, they
instead use the function in their base class.

There is another kind of initialization code in the DataModel
subclasses that could not be eliminated. That is the code that
initializes arrays to their default values when they are otherwise not
present. These are the lines that look like:

    self.dq = self.dq

These lines could not be eliminated because there is no indication in
the schema files which arrays should be initialized and which should
not.