Add collection property and to_model() function
Some datamodels are collections of submodels. The most obvious example
is ModelContainer, but there are also several Multi-Models
(MultiExposureModel, MultiSlitModel, and others.) This update provides
a consistent way to access the submodels. A new property, collection,
has been added to DataModel. It returns a list of nodes if the schema
has a single top level array. Otherwise, it raises an
AttributeError. Each node represents a submodel. Because it is a list,
iy can be looped over or accessed by index. The supporting method
to_model() has been added to ObjectNode. It converts a node to a
datamodel, but only if the node contains a meta.model_type
attribute. If it does not, it raises a ValueError. The method of the
same name in ListNode converts the entire array to a list of models.

The corresponding methods collection and to_model() in ModelContainer
allow the same code to loop over the models in a model container. The
following code examples will work for both:

```
model = dmopen(filename)
node_list = model.collection
for node in node_list:
    submodel = node.to_model()
    print(submodel.info())
submodels = model.collection.to_model()
for submodel in submodels:
    print(submodel.info())
```