Attempt to fix ResourceWarning in regression tests
Regression tests are throwing an error because when testing for a
warning message about DATAMODL not being found, it gets a
ResourceWarning about a file being left open instead. I believe the
error is caused by a fits file being left open, because the code which
schedules the fits file to be closed is after the warning
message. When a warning filter is set to convert the warning to an
error, this code is never executed. So I have moved the code that
schedules the fits file to be closed before the warning message and
modified the test so that is done outside the with block that opens
the datamodel. This should make sure the fits file is closed and get
rid of the ResourceWarning.