Support asdf flags when creating a new datamodel [#1215]
Asdf supports two flags, ignore_version_mismatch and
ignore_unrecognized_flag, when creating a new asdf file. Part of
creating a new datamodel is opening an asdf file, either natively or
as embedded in a FITS extension. This update passes these two flags
from the DataModel argument list to the call that opens or creates an
asdf file. The result is better control of the warning messages
produced by asdf.