Add validate_required_fields to DataModel (#2589)
An issue was raised during a meeting that datamodels was not correctly
validating schema items with the fits_required attribute. I did some
checking and found that fits_required is used in validation when
reading a fits file or deleting a model value. But it was not used in
validation when writing a model. The reason is that when writing the
code traverses the model, not the schema, so fields that are missing
in the model are not detected. I've added a new method,
validate_required_fields, to model_base.py that traverses the schema
and checks all fields where fits_required is true. This method is
called in save() before writing the file. It honors the usual flags,
pass_invalid_values and strict_validation.

I also added two tests to test_models.py to check the handling of
fits_required. Fits_required is not currently used in any of the
schemas, so for the present the new method has no effect.