Remove @not_under_travis and @require_crds_context pytest markers
The `@not_under_travis` marker is redundant with `@runslow` or
preferably `@pytest.mark.slow`.  And it is currently not used.

The `@require_crds_context` marker is not used at all. And in any
case is available now from `ci_watson` pytest plugin.

Also solves a deprecation warning we were seeing.