Changed the handling of zero wavelengths for NIRSpec data.
Function combine_fast_slow now calls a new function clean_wl, which returns
a copy of the wavelength array with zero or NaN values replaced with an
average.  The width of a pixel in wavelength units is computed by subtracting
the wavelengths of adjacent pixels, but now this gives better results for
pixels that are near the border between good wavelength values and NaN
values.  Before calling g_average, there's a check that the wavelength
is positive.

A check was added to verify that the wavelengths read from the fast-variation
table are increasing, and a warning is logged if they are not.

Function wl_interpolate (used when computing the fast-variation flat) now
returns the first or last element of the flat array if the wavelength is
outside the range of the wavelength array.

Function interpolate_flat (used when computing the slow-variation flat)
now only adds NO_FLAT_FIELD to the flat DQ array if the wavelength is
less than or equal to zero.

See issue #2621.