Fix bug in datamodels default value when schema contains combiner
Datamodels sets a default value when code accesses a non-existent
attribute that is mentioned in the schema. Normally this is None, but
in the case where the schema type is array, the default is a list and
when the schema type is an object, the default is a dictionary. The
code failed when a schema contained a combiner (anyOf, allOf), because
it did not search the combined subschemas. This update resolves the
problem.