Add build_docstring to schema.py
The function build_docstring creates a string from the schema file for
a datamodels class and a template. The template contains schema field
names in brackets along with other text. The schema items selected
from the schema are those that have a fits_hdu field. This update also
contains a script, schemadoc that runs build_docstring for all the
datamodels or those specified on the command line.