Check table column names against schema
This update checks table column names gainst the column names in the
schema datatype and throws a ValueError if they don't match. Matching
is case insensitive. Prviously the code only checked if the number of
columns matched. There's no option to change the error into a warning as
columns are accessed by name.