No extrapolation for NIRSpec flat field
There are several components for the NIRSpec flat field.  For each section
of the instrument (the fore-optics, the spectrograph, and the detector),
there can be both a fast variation (with wavelength) and a slow variation
component.  For a given pixel of the science data, it may be that the
wavelength of the science data is undefined, or the wavelength might be
out of range for one or more components of the flat field, or the pixel
can be flagged as bad in the reference file for a component.  In any such
case, the flat field to be applied to the data will be set to 1, and the
associated data quality array will be non-zero, such as UNDEFINED_FLAT or
NO_FLAT_FIELD.  See issue #2677.