Add a function to generate documentation from schema files
This update adds the function build_docstring to schema.py. It takes a
datamodels class and extracts field values from its schema for each
schema item associated with a fits hdu. It then inserts the field
values into a format string, which is passed as the second
argument. This update also includes a new script, schemadoc,which
calls it.