Fix for correct Read Noise in ramp segments (Issue #2790)
This change corrects for the units of read noise in the ramp segments within an integration by dividing by the gain^2. It also removes the fix for Issue #2756 since the correction is now being done at the ramp segment level and should not be applied again at the integration level.