Fixed indentation of del gain2d and the gain conversion
Based on the comments in the pull request. I corrected the indentation of the del gain_2d statement (the code worked correctly either way). I also moved the unit conversion to be on the same line as the creation of the segment variance due to read noise (and added a comment).