Fix for only two good groups in interval
This is a fix for issue #2832. When the segment only has two groups it does not get a solution. The bug is that the interval length was set to 1 instead of 2.
The tests have been updated to put back in the tests that failed. They now work with the fixes.