Update ModelContainer to accept already open associations (#2971)
* Add validate_required_fields to DataModel

An issue was raised during a meeting that datamodels was not correctly
validating schema items with the fits_required attribute. I did some
checking and found that fits_required is used in validation when
reading a fits file or deleting a model value. But it was not used in
validation when writing a model. The reason is that when writing the
code traverses the model, not the schema, so fields that are missing
in the model are not detected. I've added a new method,
validate_required_fields, to model_base.py that traverses the schema
and checks all fields where fits_required is true. This method is
called in save() before writing the file. It honors the usual flags,
pass_invalid_values and strict_validation.

I also added two tests to test_models.py to check the handling of
fits_required. Fits_required is not currently used in any of the
schemas, so for the present the new method has no effect.

* Add validate_required_fields to DataModel

An issue was raised during a meeting that datamodels was not correctly
validating schema items with the fits_required attribute. I did some
checking and found that fits_required is used in validation when
reading a fits file or deleting a model value. But it was not used in
validation when writing a model. The reason is that when writing the
code traverses the model, not the schema, so fields that are missing
in the model are not detected. I've added a new method,
validate_required_fields, to model_base.py that traverses the schema
and checks all fields where fits_required is true. This method is
called in save() before writing the file. It honors the usual flags,
pass_invalid_values and strict_validation.

I also added two tests to test_models.py to check the handling of
fits_required. Fits_required is not currently used in any of the
schemas, so for the present the new method has no effect.

* Remove changes in last commit from the master branch

* Add "import jsonschema" to test_models.py

* Add a comment to validate_required fields

* Rename configuration files [#2577]

Three configuration files were rnamed to include dashes in their
names. This was done so that the filenames agree with the association
file names. The tests have been updated to used the new configuration
file names.

* Truncate long schema validation error messages

Datamodels can issue some very long schema validation errors. This update
truncates the error message to 2000 characters if it is longer, which is the
size of an old-fashioned terminal screen (80x25).

* Wrap long line

Split a long line so that all lines are less than 80 characters.

* Undo changes meant for a branch that were accidentally committed to master

* Truncate long schema validation errors

This update truncates schema validation warnings or errors at 2000
characters, which is the size of a terminal screen (80x25). Schema
validation can create very large eror messages and these obscure the
real problems with regression test runs.

* Add line to CHANGES file about schema validation error truncation

* Add version number to CHANGE file line

* Update ModelContainer to accept already open associations

Prviously ModelContainer only accepted filenames. It's been modified
to work if its argument is an open association. The generic open
function in util.py has also been modified to accept an open
association.

* Add is_association to test if a dictionary contains an association

Associations are plain dictionaries. The new fumction looks for the
fields asn_pool and asn_id and if it finds them, says the dictionary
is an association.

* Add pushdir context manager

* Fix container bug with opening association structure

Changes:
- Fix container bug with opening association structure.
- Update test structure to match the "same folder" paradigm.

* Repair lint errors

* Updates from review comments

* Repair lint errors