Modified crds_client.py and step.py to parameterize the value of "observatory" which is passed into CRDS and previously hard coded as 'jwst' in crds_client.
This enables other telescopes to use stpipe with no customizations provided
that they perform required customizations for CRDS or a fork of CRDS.

The value of observatory is determined as follows:

1. If the new observatory parameter is set to a string,  that is used
after coercing to lower case.

2. If the new observatory parameter is omitted and a datamodel is available
in the function call,  observatory is taken from model.meta.telescope and
coerced to lower case.

3. If the observatory parameter is missing,  or the model value for telescope
is None,  observatory defaults to 'jwst'.

Two other unrelated changes are included in this commit:

1. The Step.reference_uri_to_cache_path() method is converted to a
classmethod.  On an instance, it works as before.  Called relative to a class,
it now also works, removing the need to instantiate a Step.  All-in-all, for
testers it removes the need to import crds_client just to call the related URI
translation function and decouples tests from CRDS more.

2. I removed Python-2 era code in crds_client which was used to truncate
tracebacks for routine CRDS exceptions; under Python-3 I believe the deep
tracebacks are no longer avoidable using that method.  Although this is a
reasonable change, some form of fatal error which cleanly exits would be more
appropriate, a traceback just obfuscates "couldn't find reference file", etc.