Function to convert 1-D spectrum to 2-D
Function expand_to_2d was added, to expand a 1-D background file to 2-D.
A utility function create_background was also added; this takes 1-D arrays
of wavelength and background flux, and copies those values to a table in
the appropriate format for a 1-D master background spectrum.
See issue #2308.