Address comments.
In create_master_bkg.py, I have left the data type for DQ as int32 for
the time being, because that will need to be changed in spec.schema.yaml,
and changing that would break all extract_1d regression tests.  I need to
change that schema to add a new column anyway, so I'll change the data type
for DQ at that time.