Add new NPIXELS column to output extract1d tables. (#3108)
* MultiSpec model now uses different keywords for saving time values

The schema for MultiSpecModel was modified to change the names of the
extension-header keywords for the beginning, middle, and end of an
integration.  Closes issue 2274.

* Updated the change log.

* Add new NPIXELS column to output extract1d tables.

The number of pixels that were added together to get the net count rate
is now saved in a new NPIXELS column.  This involved changes to extract_1d,
the schemas for MultiSpecModel and SpecModel, and for a utility function
in master_background.

The data type of the DQ column has been changed from int32 to uint32.

The functions in extract1d.py that do the actual spectral extraction (except
when a reference image is used) now compute and save the results as float64
instead of float32.

This closes issue #3107.