Modify combine_1d for combining background data
For the case that the FLUX column of the input extract1d tables is actually
background, the combine_1d step now scales the values read from the FLUX
column by dividing by the NPIXELS column.  Also in this case, NPIXELS is
included as a factor in the weight (if there's more data, the weight should
be higher).

A new command-line parameter `background` was added, to allow specifying
that the data are actually background rather than target flux.

The default for weighting the input tables has been changed from the
integration time to the exposure time; see parameter `exptime_key`.
Closes issue #3163.