Changes to the combine_1d step (#3220)
* Modify combine_1d for combining background data

For the case that the FLUX column of the input extract1d tables is actually
background, the combine_1d step now scales the values read from the FLUX
column by dividing by the NPIXELS column.  Also in this case, NPIXELS is
included as a factor in the weight (if there's more data, the weight should
be higher).

A new command-line parameter `background` was added, to allow specifying
that the data are actually background rather than target flux.

The default for weighting the input tables has been changed from the
integration time to the exposure time; see parameter `exptime_key`.
Closes issue #3163.

* Don't log a message unless it's relevant.

* Changes to the combine_1d step

The combine_1d step can be invoked by calling combine1d.do_correction.  The
first argument is now an open jwst.datamodels model, rather than the name
of the input file or association.  That model can be either a ModelContainer
or a MultiSpecModel.

Closes #3208.
Closes #3210.

* do_correction changed to do_combine.

* Change function name in combine_1d

The function to call combine_1d without using the step interface has been
changed to combine_1d_spectra.  References to the `interpolation` parameter
have mostly been removed, since it isn't used and it was not clear what it
referred to.

Closes #3208.
Closes #3210.

* A few more changes to combine_1d