Allow reference image for IFU to be either 2-D or 3-D
Since the width of a PSF will vary with wavelength, the extract_1d step
can use either a 2-D reference image (same shape as one plane of the IFU
cube) or a 3-D reference image (same shape as the entire IFU cube).

Resolves #3121.
Closes #3121.
Resolves #3213.
Closes #3213.

When using a reference image for non-IFU data, the background regions are
now smoothed after scaling the background to match the number of pixels in
the source extraction region.

Resolves #3260.
Closes #3260.