Combine1DStep needed arrays of right ascension and declination
The combine_1d step failed in the call to wcs.invert because two of the
arguments (right ascension and declination) were scalars, but the wavelength
was an array.  Right ascension and declination are now stored as arrays,
even though their values are constant.

Another bug was that in the files for the test case, the net column was all
zeros (because the spectra had been extracted from flux-calibrated data).
The flux was being weighted by the sensitivity (so that the quantity that
was accumulated would be in units of counts), but the sensitivity was gotten
as the ratio of net to flux; therefore, the weight was being set to zero.
A test has been added to disable including the sensitivity in the weight for
flux if the net is zero.

Closes #3270.