Use target coordinates to shift extraction location
The right ascension and declination of the target are now gotten from
keywords TARG_RA and TARG_DEC, and the inverse WCS is used to find the
location of that celestial point in the input image.  For a 2-D input
image (or a stack of such images), that point is assumed to be the
location of the spectrum in the cross-dispersion direction.  For an IFU
cube, the forward WCS is evaluated at each pixel in one plane to compute
the right ascension and declination (and wavelength), and the location of
the target is taken to be the pixel with celestial coordinates that are
closest to TARG_RA and TARG_DEC.  The extraction location will actually
be shifted to the computed point, except for extended sources, WFSS
modes, or (currently) NIRSpec fixed-slit or MOS.

For IFU, the computation of the areas of the source and background regions
has been changed.  Instead of calling aperture.area() or annulus.area(),
a temporary array filled with ones is created, and the aperture and annulus
are applied to that array.  The 'aperture_sum' in the photometry table
is then the actual area, i.e. accounting for the possibility that the
aperture or annulus extends beyond the borders of the IFU image.

See JP-597 and JP-644.

Closes #3284.  Closes #3361.