x_center and y_center were referenced before being defined
For IFU data, x_center and y_center were only assigned values if the target
was not an extended source.  For an extended source, all relevant parameters
are now assigned values, and the values are set to extract over the entire
input image (ignoring the reference file, if any), as specified in recent
requirements.

A try/except block was added to check for the case where a WCS does not have
an inverse.

The code that checks for the existence of a RELSENS extension (this code will
soon be removed, but not yet), there is a check on the length of the relsens
attribute, and this fails with "TypeError: 'int' object is not callable".
The RELSENS extension is no longer needed anyway, so this check has been
deleted.

Closes #3381.