Modify extract_1d and combine_1d for change in units
The photom step now applies the photometric correction, and extract_1d and
combine_1d have been updated to accept an input file in units of surface
brightness.  extract_1d no longer checks for a PHOTOM extension, and the
output table has new columns SURF_BRIGHT and SB_ERROR rather than NET and
NERROR.  In addition, the BACKGOUND column has been divided by NPIXELS, so
the units will be surface brightness.  Previously, the BACKGROUND column
contained the values that were subtracted from the gross count rate to get
the net count rate.  The ref_file attribute has been removed from the
step class, Extract1dStep; ref_file is now a local variable.

The background parameter has been removed from the combine_1d step.  This
is no longer needed, since extract_1d divides the background by npixels.
combine_1d reads the SURF_BRIGHT and SB_ERROR columns (if they exist) from
the input file(s), ignoring NET.  The weight no longer includes NPIXELS
for averaging background, under the assumption that the input files will be
similar to each other if it is the background that is of interest.  When
averaging the FLUX and its error, the sensitivity is no longer included
as a factor in the weight, because determining the sensitivity required
the net count rate, which is no longer available.

Closes #3295.
Closes #3301.
Closes #3392.