Apply barshadow, pathloss, and photom correction to VAR_FLAT
The correction factors (squared) are applied to the VAR_POISSON and
VAR_RNOISE variances in the barshadow, pathloss, and photom steps.  These
correction factors (also squared) are now also applied to the VAR_FLAT
variance array.

Note that this PR does not close issue #3414, because there are other steps
that can follow flat_field.