For Jira JP-741 some members of a coron3 association should not be processed/prefetched.
Modified step.py to expose a .prefetch() method which is called just prior to
.process() IFF .prefetch_references == True.   This is the default.

If .prefetch_references == False,  .prefetch() can still be called explicitly
within .process() if needed,  as is done here for calwebb_coron3.

To improve customizability, changed Pipeline.reference_file_types, once defined
in __init__() as an attribute, into a @property which is deferred until needed.
This presents the possibility of overloading the property to redefine the
reference_file_types used by a pipeline.  Since reference_file_types is
already a class attribute in Step, the ability to override for Step already
existed, even as an instance @property.